refactored the navigation handling and generated test to check that it navigates properly to the detail screen on article press from the feed screen
