+++ src/ui/Detail/Detail.controller.tsx
@@ -1,4 +1,4 @@
-import { DetailProps } from '../NavigationTypes'
+import { DetailProps } from '../_navigation/NavigationTypes'
 import { Detail } from './Detail'
 
 export const DetailController = (props: DetailProps) => <Detail />
