+++ src/ui/Detail/Detail.tsx
@@ -1,7 +1,7 @@
 import { View, Text } from 'react-native'
 
 export const Detail = () => (
-  <View>
+  <View testID="detailScreen">
     <Text>This is the detail page</Text>
   </View>
 )
