+++ src/ui/Feed/Feed.controller.tsx
@@ -1,5 +1,5 @@
 import { aSetOfArticles, sortArticles } from '../../core/domain/model/Feed/Feed.model'
-import { FeedProps } from '../NavigationTypes'
+import { FeedProps } from '../_navigation/NavigationTypes'
 import { Feed } from './Feed'
 
 const FakeData = sortArticles(aSetOfArticles(30))
